requirements.txt
# FastAPI & ASGI Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20

# RabbitMQ
aio-pika==9.5.4

# Redis
redis[hiredis]==5.2.1

# HTTP Client (Sarvam AI + Testing)
httpx==0.28.1

# Monitoring
prometheus-client==0.21.1
prometheus-fastapi-instrumentator==7.0.2

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0

# Machine Learning
scikit-learn==1.4.2
joblib==1.4.2
pandas==2.2.2
numpy==1.26.4

# Utilities
python-dotenv==1.0.1
